Game Resources Refactor First Commit
Game Resources Refactor First Commit.
10/8 Done. Issue #630
