+++ Assets/Source/Agents/AgentCreationApi.cs
@@ -184,6 +184,100 @@ namespace Agent
         {
             CurrentIndex = -1;
         }
+
+
+
+        public void InitializeResources()
+        {
+
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.Player);
+            GameState.AgentCreationApi.SetName("player");
+            GameState.AgentCreationApi.SetMovement(10f, 3.5f, 2);
+            GameState.AgentCreationApi.SetHealth(300.0f);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.5f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(-0.35f, 0.0f), new Vec2f(0.75f, 2.8f));
+            GameState.AgentCreationApi.End();
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.Agent);
+            GameState.AgentCreationApi.SetName("agent");
+            GameState.AgentCreationApi.SetMovement(5f, 3.5f, 1);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.5f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(0.25f, 0.0f), new Vec2f(0.5f, 1.5f));
+            GameState.AgentCreationApi.SetStartingAnimation((int)Animation.AnimationType.CharacterMoveLeft);
+            GameState.AgentCreationApi.End();
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.Slime);
+            GameState.AgentCreationApi.SetName("Slime");
+            GameState.AgentCreationApi.SetMovement(5f, 3.5f, 1);
+            GameState.AgentCreationApi.SetDropTableID(Enums.LootTableType.SlimeEnemyDrop, Enums.LootTableType.SlimeEnemyDrop);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.0f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(0.125f, 0.0f), new Vec2f(0.75f, 0.5f));
+            GameState.AgentCreationApi.SetStartingAnimation((int)Animation.AnimationType.SlimeMoveLeft);
+            GameState.AgentCreationApi.SetEnemyBehaviour(Agent.EnemyBehaviour.Slime);
+            GameState.AgentCreationApi.SetDetectionRadius(4.0f);
+            GameState.AgentCreationApi.SetHealth(100.0f);
+            GameState.AgentCreationApi.SetAttackCooldown(0.8f);
+            GameState.AgentCreationApi.End();
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.FlyingSlime);
+            GameState.AgentCreationApi.SetName("Flying Slime");
+            GameState.AgentCreationApi.SetFlyingMovement(3.0f);
+            GameState.AgentCreationApi.SetDropTableID(Enums.LootTableType.SlimeEnemyDrop, Enums.LootTableType.SlimeEnemyDrop);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.0f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(0.125f, 0.0f), new Vec2f(0.75f, 0.5f));
+            GameState.AgentCreationApi.SetStartingAnimation((int)Animation.AnimationType.SlimeMoveLeft);
+            GameState.AgentCreationApi.SetEnemyBehaviour(Agent.EnemyBehaviour.Slime);
+            GameState.AgentCreationApi.SetDetectionRadius(4.0f);
+            GameState.AgentCreationApi.SetHealth(100.0f);
+            GameState.AgentCreationApi.SetAttackCooldown(0.8f);
+            GameState.AgentCreationApi.End();
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.EnemySwordman);
+            GameState.AgentCreationApi.SetName("enemy-swordman");
+            GameState.AgentCreationApi.SetMovement(3f, 3.5f, 2);
+            GameState.AgentCreationApi.SetDropTableID(Enums.LootTableType.SlimeEnemyDrop, Enums.LootTableType.SlimeEnemyDrop);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.5f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(-0.25f, 0.0f), new Vec2f(0.75f, 2.5f));
+            GameState.AgentCreationApi.SetEnemyBehaviour(Agent.EnemyBehaviour.Swordman);
+            GameState.AgentCreationApi.SetDetectionRadius(16.0f);
+            GameState.AgentCreationApi.SetHealth(100.0f);
+            GameState.AgentCreationApi.End();
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.EnemyGunner);
+            GameState.AgentCreationApi.SetName("enemy-gunner");
+            GameState.AgentCreationApi.SetMovement(3f, 3.5f, 2);
+            GameState.AgentCreationApi.SetDropTableID(Enums.LootTableType.SlimeEnemyDrop, Enums.LootTableType.SlimeEnemyDrop);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.5f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(-0.25f, 0.0f), new Vec2f(0.75f, 2.5f));
+            GameState.AgentCreationApi.SetEnemyBehaviour(Agent.EnemyBehaviour.Gunner);
+            GameState.AgentCreationApi.SetDetectionRadius(24.0f);
+            GameState.AgentCreationApi.SetHealth(100.0f);
+            GameState.AgentCreationApi.End();
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.EnemyInsect);
+            GameState.AgentCreationApi.SetName("enemy-insect");
+            GameState.AgentCreationApi.SetMovement(3f, 3.5f, 2);
+            GameState.AgentCreationApi.SetDropTableID(Enums.LootTableType.SlimeEnemyDrop, Enums.LootTableType.SlimeEnemyDrop);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.5f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(-0.25f, 0.0f), new Vec2f(0.75f, 2.5f));
+            GameState.AgentCreationApi.SetEnemyBehaviour(Agent.EnemyBehaviour.Insect);
+            GameState.AgentCreationApi.SetDetectionRadius(16.0f);
+            GameState.AgentCreationApi.SetHealth(100.0f);
+            GameState.AgentCreationApi.End();
+
+
+            GameState.AgentCreationApi.Create((int)Enums.AgentType.EnemyHeavy);
+            GameState.AgentCreationApi.SetName("enemy-insect-heavy");
+            GameState.AgentCreationApi.SetMovement(3f, 3.5f, 2);
+            GameState.AgentCreationApi.SetDropTableID(Enums.LootTableType.SlimeEnemyDrop, Enums.LootTableType.SlimeEnemyDrop);
+            GameState.AgentCreationApi.SetSpriteSize(new Vec2f(1.0f, 1.5f));
+            GameState.AgentCreationApi.SetCollisionBox(new Vec2f(-0.25f, 0.0f), new Vec2f(0.75f, 2.5f));
+            GameState.AgentCreationApi.SetEnemyBehaviour(Agent.EnemyBehaviour.Insect);
+            GameState.AgentCreationApi.SetDetectionRadius(16.0f);
+            GameState.AgentCreationApi.SetHealth(100.0f);
+            GameState.AgentCreationApi.End();
+        }
     }
 
 }
