+++ Assets/Source/TGen/Rendering/RenderGridOverlay.cs
@@ -87,5 +87,39 @@ namespace TGen
             mesh.SetTriangles(triangles, 0);
         }
 
+        public int TGenBlockSpriteSheet;
+        public int[] TGenIsotypeSprites;
+
+        public void InitializeResources()
+        {
+            TGenBlockSpriteSheet = GameState.SpriteLoader.GetSpriteSheetID("Assets\\StreamingAssets\\Tiles\\Blocks\\Test\\testBlocks.png", 32, 32);
+
+            var emptySprite = GameState.SpriteLoader.GetSpriteSheetID("Assets\\StreamingAssets\\Tiles\\TileCollision\\EmptyBlock.png", 32, 32);
+
+            int tileCount = 42;
+
+            TGenIsotypeSprites = new int[tileCount];
+
+            TGenIsotypeSprites[0] = GameState.SpriteAtlasManager.CopySpriteToAtlas(emptySprite, 0, 0, Enums.AtlasType.TGen);
+
+            TGenIsotypeSprites[1] = GameState.SpriteAtlasManager.CopySpriteToAtlas(TGenBlockSpriteSheet, 1, 1, Enums.AtlasType.TGen);
+
+            var row = 3;
+            var column = 1;
+
+            for (int i = 2; i < tileCount; i++)
+            {
+                TGenIsotypeSprites[i] = GameState.SpriteAtlasManager.CopySpriteToAtlas(TGenBlockSpriteSheet, column, row, Enums.AtlasType.TGen);
+
+                column += 2;
+
+                if (column > 8)
+                {
+                    column = 1;
+                    row += 2;
+                }
+            }
+        }
+
     }
 }
